make datetime picker to be genericdatetimepicker
